# GC: optimize sweeping