Moves recovering small pages into sweep() and optimizes sweep by precalculating some common scenarios.
With the extended vdparser-test on win64:
* master: GC summary: 1355 MB,   13 GC 1443 ms, mark 1089 ms, sweep **199** ms, recover **150** ms, Pauses 1092 ms <  410 ms
* this PR: GC summary: 1355 MB,   13 GC 1226 ms, mark 1082 ms, sweep  **16** ms, recover **124** ms, Pauses 1085 ms <  408 ms